Updates dependencies and cart store logic
Updates several dependencies to their latest versions, including Prisma, better-auth, lucide-react, and zod.

Refactors the cart store to use CartItemDTO, which includes the product and quantity, instead of including product data inline within the item.

Introduces `syncItems` to the cart store to hydrate the store from an array of `CartItemDTO` which is loaded on application initialization.