Adds set default address functionality
Implements the ability for a user to set a default shipping address.

This change introduces a new API endpoint and corresponding logic to
allow users to designate one of their addresses as the default. This
simplifies the checkout process and improves user experience by
automatically pre-selecting the default address.

Addresses the need to have a single, preferred shipping destination.